Creation of a system that get the player performance's data and store in a csv file

+++ Assets/Scripts/GameController.cs
@@ -10,8 +10,8 @@ public class GameController : MonoBehaviour
     private int currentLevel;
     [SerializeField] private List<LevelData> levels;
     private MenuController menuController;
-    [SerializeField] private int difficultLevel;
-    public int DifficultLevel
+    [SerializeField] private float difficultLevel;
+    public float DifficultLevel
     {
         get {return difficultLevel; }
         set {difficultLevel = value; }
